update method call description
- add region related configuration description
